### 2025-11-16 #9419 @editorialbot
[PRE REVIEW]: ExonViz: Transcript visualization made easy
**Submitting author:** @Redmar-van-den-Berg (<a href="http://orcid.org/0000-0002-6590-9499">Redmar R. van den Berg</a>)
**Repository:** https://github.com/DCRT-LUMC/exonviz
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.14
