+++ .github/workflows/ci.yml
@@ -35,11 +35,11 @@ jobs:
     strategy:
       matrix:
         python-version:
-          - "3.9"
           - "3.10"
           - "3.11"
           - "3.12"
           - "3.13"
+          - "3.14"
     needs: sanity
     steps:
       - uses: actions/checkout@v3
