### 2025-12-03 f07b75b
Update typing syntax
