+++ src/exonviz/mutalyzer.py
@@ -1,4 +1,4 @@
-from typing import Dict, Any, List, Tuple, cast
+from typing import Any, cast
 import urllib.request
 from urllib.error import HTTPError
 import json
@@ -15,7 +15,7 @@ import logging
 logging.basicConfig()
 log = logging.getLogger(__name__)
 
-Range = Tuple[int, int]
+Range = tuple[int, int]
 
 
 def parse_error_payload(error: HTTPError) -> str:
@@ -28,7 +28,7 @@ def parse_error_payload(error: HTTPError) -> str:
     return msg if msg else str(error)
 
 
-def fetch_exons(transcript: str) -> Dict[str, Any]:
+def fetch_exons(transcript: str) -> dict[str, Any]:
     """Fetch transcript information from mutalyzer"""
 
     url = f"https://mutalyzer.nl/api/normalize/{transcript}"
@@ -44,11 +44,11 @@ def fetch_exons(transcript: str) -> Dict[str, Any]:
     if "selector_short" not in js:
         msg = f"No exons found for {transcript} (is it a genomic variant?)"
         raise RuntimeError(msg)
-    selector: Dict[str, Any] = js["selector_short"]
+    selector: dict[str, Any] = js["selector_short"]
     return selector
 
 
-def fetch_variants(transcript: str) -> Dict[str, Any]:
+def fetch_variants(transcript: str) -> dict[str, Any]:
     """Fetch variant view from mutalyzer"""
 
     url = f"https://mutalyzer.nl/api/view_variants/{transcript}"
@@ -59,11 +59,11 @@ def fetch_variants(transcript: str) -> Dict[str, Any]:
         msg = parse_error_payload(e)
         raise RuntimeError(msg)
     else:
-        js: Dict[str, Any] = json.loads(response.read())
+        js: dict[str, Any] = json.loads(response.read())
     return js
 
 
-def convert_mutalyzer_range(start: str, end: str) -> Tuple[int, int]:
+def convert_mutalyzer_range(start: str, end: str) -> tuple[int, int]:
     """Convert a mutalyzer range to 0-based coordinates"""
 
     if is_reverse(start, end):
@@ -77,7 +77,7 @@ def is_reverse(start: str, end: str) -> bool:
     return int(start) > int(end)
 
 
-def convert_exon_positions(positions: List[List[str]]) -> List[Tuple[int, int]]:
+def convert_exon_positions(positions: list[list[str]]) -> list[tuple[int, int]]:
     """Convert exon positions from Mutalyzer to a list of Ranges
 
     This function also accounts for reverse transcripts
@@ -106,7 +106,7 @@ def make_coding(exon: Range, coding_region: Range, start_phase: int) -> Coding:
     )
 
 
-def exon_variant(exons: List[List[str]], variant: Dict[str, Any]) -> bool:
+def exon_variant(exons: list[list[str]], variant: dict[str, Any]) -> bool:
     """Determine if a given variant falls in an exon"""
     reverse = is_reverse(exons[0][0], exons[0][1])
 
@@ -120,8 +120,8 @@ def exon_variant(exons: List[List[str]], variant: Dict[str, Any]) -> bool:
 
 
 def variants_outside_exons(
-    exons: List[List[str]], payload: List[Dict[str, Any]]
-) -> List[str]:
+    exons: list[list[str]], payload: list[dict[str, Any]]
+) -> list[str]:
     """Determine description of variants that fall outside the exons"""
     # Exclude the regions that flank the variants
     variants = [x for x in payload if x["type"] == "variant"]
@@ -134,8 +134,8 @@ def variants_outside_exons(
 
 
[…]
