### 2025-12-03 09750bc
Update typing syntax
