+++ src/exonviz/exon.py
@@ -1,25 +1,10 @@
 from dataclasses import dataclass
-from typing import (
-    Any,
-    List,
-    Optional,
-    Sequence,
-    Dict,
-    Tuple,
-    no_type_check,
-    Union,
-    cast,
-)
+from typing import Any, Sequence, no_type_check
 import copy
 import math
 from decimal import Decimal, ROUND_UP
 
-import sys
-
-if sys.version_info >= (3, 10):
-    from typing import TypeAlias
-else:
-    from typing_extensions import TypeAlias
+from typing import TypeAlias
 
 from _io import TextIOWrapper
 from svg import Rect, Polygon, Text, Style, Circle
@@ -30,7 +15,7 @@ import logging
 
 logging.basicConfig(level="DEBUG")
 log = logging.getLogger(__name__)
-Element: TypeAlias = Union[Circle, Rect, Polygon, Text, Style]
+Element = Circle | Rect | Polygon | Text | Style
 
 
 @dataclass()
@@ -132,8 +117,8 @@ class Exon:
     def __init__(
         self,
         size: int,
-        coding: Optional[Coding] = None,
-        variants: Optional[Sequence[Variant]] = None,
+        coding: Coding | None = None,
+        variants: Sequence[Variant] | None = None,
         name: str = "",
         color: str = "#4C72B7",
     ) -> None:
@@ -209,12 +194,12 @@ class Exon:
         x: float = 0,
         y: float = 0,
         variant_shape: str = "pin",
-    ) -> List[Element]:
+    ) -> list[Element]:
         """Draw the Exon, in SVG format
 
         Returns a list of SVG elements
         """
-        elements: List[Any] = list()
+        elements: list[Element] = list()
 
         # If the coding size is not the entire exon
         if self.coding.size != self.size:
@@ -302,7 +287,7 @@ class Exon:
             raise ValueError(msg)
 
         # fmt: off
-        start: List[List[float]] = [
+        start: list[list[float]] = [
             [ # Square
                 cx, y + height,
                 cx, y
@@ -319,7 +304,7 @@ class Exon:
             ]
         ]
 
-        end: List[List[float]] = [
+        end: list[list[float]] = [
             [ # Square
                 cx + size, y,
                 cx + size, y + height,
@@ -353,7 +338,7 @@ class Exon:
         shape: str = "bar",
     ) -> Sequence[Element]:
         """Draw Variants for the Exon"""
-        elements: List[Element] = list()
+        elements: list[Element] = list()
         c_size = 0.25 * height
         for variant in self.variants:
             if shape == "pin":
@@ -434,7 +419,7 @@ class Exon:
             variants=new_variants,
         )
 
-    def valid_splits(self, height: float = 20, scale: float = 1) -> List[Range]:
+    def valid_splits(self, height: float = 20, scale: float = 1) -> list[Range]:
         """Determine which splits of this exon can be drawn"""
 
         def meaningfull(split: Range) -> bool:
@@ -497,14 +482,14 @@ class Exon:
         return sep.join(map(str, records))
 
 
-def _pick_split(splits: List[Range], page_size: int) -> int:
+def _pick_split(splits: list[Range], page_size: int) -> int:
     """
     Pick a legal split from the allowed splits and the page size
 
     For now, we just pick the biggest possible split
     """
     # Intersect each valid split with the page
-    splits_that_fit: List[Range] = list()
+    splits_that_fit: list[Range] = list()
     page = (0, page_size + 1)
     for split in splits:
[…]

[…]
