### 2025-12-03 471d83e
Install ExonViz before building the docs
