+++ .readthedocs.yaml
@@ -31,3 +31,6 @@ formats:
 python:
    install:
    - requirements: docs/requirements.txt
+   # Install the python package before building the docs
+    - method: pip
+      path: .
