+++ .readthedocs.yaml
@@ -32,5 +32,5 @@ python:
    install:
    - requirements: docs/requirements.txt
    # Install the python package before building the docs
-    - method: pip
-      path: .
+   - method: pip
+     path: .
