### 2025-12-04 49b67b5
Add function to convert c. variants into tuples for the crossmapper
