### 2025-12-04 ead6b5c
Add function to convert c. variants into tuples for the crossmapper
