### 2025-12-08 3f9927c
Convert cdot variants to coordinate positions with the crossmapper
