### 2025-12-08 dc4a346
Take that start for variants which are ranges
