### 2025-12-08 62293f6
Ensure variants are in order, refactor
