### 2025-12-08 2b1eb71
Move sorting of variants to check input function
