+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ Changelog
 .. This document is user facing. Please word the changes in such a way
 .. that users understand how the changes affect the new version.
 
-v0.2.16-dev
+v0.2.16
 -----------
 + Add API documentation
 + Fix a bug with variant positions in transcripts with exons
