### 2025-12-09 5e314f2
Fix bug in sort_variants that could lead to loss of a variant
