### 2025-12-09 ac31e7f
Fix bug in sort_variants that could lead to loss of a variant
