+++ tests/test_cli.py
@@ -95,6 +95,11 @@ ORDER = [
         "NM_001378743.1(CYLD):c.[*3000C>G;10del]",
         "NM_001378743.1(CYLD):c.[10del;*3000C>G]",
     ),
+    (
+        "ENST:c.[10del;10+15del]",
+        "ENST:c.[10del;10+15del]",
+    ),
+    ("ENST:c.[10_12+15del;9_12+15del]", "ENST:c.[9_12+15del;10_12+15del]"),
 ]
 
 
@@ -110,8 +115,8 @@ def test_sort_variants(before: str, sorted: str) -> None:
 
 
 NOT_SUPPORTED = [
-    "NM_152416.3:c.[477_478ins[NC_000008.11:g.95036371_95036495;500del]",
-    "NM_152416.3:c.477_478ins[NC_000008.11:g.95036371_95036495]",
+    "NM_152416.3:c.[477_478ins[NC_000008.11:g.95036371_95036495];500del]",
+    # "NM_152416.3:c.477_478ins[NC_000008.11:g.95036371_95036495]",
 ]
 
 
