### 2025-12-11 7ef183a
Automatically add changelog to release notes
