### 2025-12-11 3c3bfcd
Automatically add changelog to release notes
