+++ CHANGELOG.rst
@@ -7,6 +7,10 @@ Changelog
 .. This document is user facing. Please word the changes in such a way
 .. that users understand how the changes affect the new version.
 
+-----------
+v0.2.18-dev
+-----------
+
 -------
 v0.2.17
 -------
