### 2026-01-23 b12ba3c
Add section on research impact, software design
