### 2026-01-23 8c493c9
Cite guidelines from Cheerie
