### 2026-01-23 e374a62
Move website stats to the impact statement section
