+++ paper.md
@@ -40,12 +40,12 @@ Transcripts of a gene contain one or more **exons**, which encode the
 functional parts of the transcript, and **introns**, which are removed in a
 process called **splicing**. A single gene typically encodes multiple
 **transcripts** by including different exons. Protein coding genes include one
-or more coding exons, which encode the protein using three-letter sequences
+or more coding exons, which encode a protein using three-letter sequences
 called **codons**. It is important to realize that exon boundaries and codon
-boundaries do not coincide, *i.e.* a codon can span two exons. If the codon
+boundaries do not coincide, *i.e.* single a codon can span two exons. If the codon
 boundaries of adjacent exons are not aligned, an often detrimental **frame
 shift** is introduced. Taking exon boundary reading frames, which we will call
-**exon boundary frames** into account is crucial when considering the effect of
+**exon boundary frames**, into account is crucial when considering the effect of
 mutations and when designing genetic therapies.
 
 `ExonViz` is a Python package and web application that creates biologically
@@ -61,7 +61,7 @@ the end of the transcript.\label{abstract}](docs/figures/abstract.svg)
 Visualization of transcripts, including features like exon boundary frames,
 coding and non coding regions is important within the field of clinical and
 human genetics [@Walker2023]. Illustrating the exon structure and the location
-of variants is common practice, especially when new genes or transcripts have
+of variants is common practice, especially when new genes, variants or transcripts have
 been discovered. These illustrations are also used to assess potential genetic
 treatment options (e.g., canonical exon skipping), in teaching settings, in
 diagnostics, to identify mutational hotspots and for genetic counseling. To
@@ -73,23 +73,23 @@ different transcript isoforms (genepainter [@Muhlhausen2015]), or visualize
 variants (Variant View [@Ferstay2013]). However, none of these tools can draw
 exon boundary frames. Knowledge about the exon boundary frames aids in the
 assessment of the pathogenicity of genetic variants using the ACMG-AMP
-guidelines [@Richards2015] when evaluating exon spanning deletions
+guidelines [@Richards2015], when evaluating exon spanning deletions
 [@Cheerie2025] and when interpreting the effects of splice altering variants
-[@Walker2023]. Creating transcript visualizations must be quick and easy if
-they are to be utilized in clinical and day to day settings, rather than to
+[@Walker2023]. Creating transcript visualizations must be quick and easy
+to be utilized in clinical and day to day settings, rather than to
 create a bespoke figure for a manuscript or presentation.
 
-To our knowledge, there currently are no easily usable tools which allow the
+To our knowledge there are currently no easily usable tools which allow the
 user to draw all features required for a comprehensive overview of a
 transcript’s structure and the location of variants of interest.
 
 # Software design
 ExonViz is written in Python 3, its web interface is build using Flask. To
-avoid the complexities of supporting various different transcript definitions,
-ExonViz uses the public Mutalyzer API [@Lefter2021] to fetch transcript
-annotations. This gives ExonViz access to all transcripts defined in the RefSeq
-[@OLeary2016] and Ensembl [@Harrison2024] databases across many species,
-ranging from human and mouse to fruit fly and coelacanth.
+avoid the complexities of retrieving and proccessing various different
+transcript definitions, ExonViz uses the public Mutalyzer API [@Lefter2021] to
+fetch transcript annotations. This gives ExonViz access to all transcripts
+defined in the RefSeq [@OLeary2016] and Ensembl [@Harrison2024] databases
+across many species, ranging from human and mouse to fruit fly and coelacanth.
 
 Reverse strand transcript are inverted so that ExonViz always visualizes
 transcripts in their forward orientation. This avoids the complications that
[…]
