### 2026-01-29 be1c6b9
Add section on the state of the field
