+++ paper.bib
@@ -48,6 +48,16 @@
   year={2015},
   publisher={Oxford University Press}
 }
+@article{Reese2021,
+  title={Swan: a library for the analysis and visualization of long-read transcriptomes},
+  author={Reese, Fairlie and Mortazavi, Ali},
+  journal={Bioinformatics},
+  volume={37},
+  number={9},
+  pages={1322--1323},
+  year={2021},
+  publisher={Oxford University Press}
+}
 @article{Richards2015,
   title={Standards and guidelines for the interpretation of sequence variants: a joint consensus recommendation of the American College of Medical Genetics and Genomics and the Association for Molecular Pathology},
   author={Richards, Sue and Aziz, Nazneen and Bale, Sherri and Bick, David and Das, Soma and Gastier-Foster, Julie and Grody, Wayne W and Hegde, Madhuri and Lyon, Elaine and Spector, Elaine and others},
