### 2026-03-09 #9419 @borgesaugusto
Hi! 👋 @duncanjunegong and @jaybee84, would you be willing to review this submission? We do our reviews here on GitHub, following a [checklist  ](https://joss.readthedocs.io/en/latest/review_criteria.html)

Thank you!
