### 2026-04-02 #9419 @biermanr
@borgesaugusto I'm willing to review this submission for JOSS. This is my first time as a reviewer
