### 2026-04-03 #9419 @borgesaugusto
Hi! 👋 @zhengxinchang and @flekschas, would you be willing to review this submission for JOSS? We do our reviews here on GitHub, following a checklist (with these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html )

Thank you!
