+++ tox.ini
@@ -9,7 +9,7 @@ envlist = py3
 description = Run py- and doctsts
 deps = pytest
        mypy
-commands = pytest
+commands = pytest tests
 
 [testenv:formatting]
 description = Check code formatting using black 
@@ -22,6 +22,7 @@ deps = mypy
        pytest
        svg.py
        types-setuptools
+       flask
 commands = mypy --strict --ignore-missing-imports src tests
 
 [testenv:coverage]
@@ -31,7 +32,7 @@ deps = coverage
        svg.py
        mypy
 commands =
-  coverage run -m pytest
+  coverage run -m pytest tests
   coverage html
   coverage xml
 
