### 2026-04-30 #9419 @samuelortion
@borgesaugusto I am willing to review this submission. This would be my first review for JOSS.
