### 2026-04-30 #10485 @editorialbot
[REVIEW]: ExonViz: Transcript visualization made easy
**Submitting author:** @Redmar-van-den-Berg (<a href="http://orcid.org/0000-0002-6590-9499">Redmar R. van den Berg</a>)
**Repository:** https://github.com/DCRT-LUMC/exonviz
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v0.2.18
