### 2026-05-01 30580cf
Pin setuptools to a version before 82
The mutalyzer crossmapper uses pkg_resources, which has been deprecated
in version 82.
