+++ tox.ini
@@ -13,7 +13,7 @@ commands = pytest tests
 
 [testenv:formatting]
 description = Check code formatting using black 
-deps = black
+deps = black>=25,<26
 commands = black --check src tests
 
 [testenv:typing]
