### 2026-05-01 5bacc42
Update svg.py to 1.10
