### 2026-05-18 a806620
Rename webserver to website, update documentation
