+++ setup.py
@@ -78,7 +78,7 @@ setup(
         # eg: 'aspectlib==1.1.1', 'six>=1.7',
     ],
     extras_require={
-        "webserver": ["flask"],
+        "website": ["flask"],
     },
     setup_requires=[
         "pytest-runner",
@@ -86,7 +86,7 @@ setup(
     entry_points={
         "console_scripts": [
             "exonviz=exonviz.cli:main",
-            "exonviz-webserver=exonviz.app:main",
+            "exonviz-website=exonviz.app:main",
         ]
     },
 )
