+++ setup.py
@@ -27,7 +27,7 @@ def read(*names, **kwargs):
 
 setup(
     name="exonviz",
-    version="0.2.18-dev",
+    version="0.2.18",
     license="AGPL-3.0",
     description="Visualise exons and their reading frames",
     long_description=long_description,
