+++ CHANGELOG.rst
@@ -7,16 +7,13 @@ Changelog
 .. This document is user facing. Please word the changes in such a way
 .. that users understand how the changes affect the new version.
 
------------
-v0.2.19-dev
------------
-+ Fix bug in generation of release notes
-+ Fix link to changelog 
-
 -------
 v0.2.18
 -------
 + Add ability to run the ExonViz website using ``exonviz-website``
++ Fix bug in generation of release notes
++ Fix link to changelog
++ Add additional checks for automated release
 
 -------
 v0.2.17
