+++ .github/workflows/release.yml
@@ -14,8 +14,12 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
+
       - name: Get shortened version of release tag
         run: echo "RELEASE_VERSION=${GITHUB_REF#refs/*/}" >> $GITHUB_ENV
+
+      - name: Check that the version does not have dev in the name
+        run: echo ${RELEASE_VERSION} | grep -v dev
 
       - name: Check release version is present in every file
         run: |
