+++ paper.bib
@@ -1,7 +1,8 @@
 @article{Alasoo2017,
   title={wiggleplotr: Make read coverage plots from BigWig files},
   author={Alasoo, Kaur},
-  journal={Bioconductor. doi},
+  journal={Bioconductor},
+  doi={10.18129/B9.bioc.wiggleplotr},
   volume={10},
   pages={B9},
   year={2017}
@@ -52,6 +53,7 @@
   title={Swan: a library for the analysis and visualization of long-read transcriptomes},
   author={Reese, Fairlie and Mortazavi, Ali},
   journal={Bioinformatics},
+  doi={10.1093/bioinformatics/btaa836}
   volume={37},
   number={9},
   pages={1322--1323},
