### 2026-05-18 4657eeb
Fix typo in bibtex file
