+++ paper.bib
@@ -53,7 +53,7 @@
   title={Swan: a library for the analysis and visualization of long-read transcriptomes},
   author={Reese, Fairlie and Mortazavi, Ali},
   journal={Bioinformatics},
-  doi={10.1093/bioinformatics/btaa836}
+  doi={10.1093/bioinformatics/btaa836},
   volume={37},
   number={9},
   pages={1322--1323},
