### 2026-06-15 e5b33a7
Partial rewrite of the software design
