### 2026-06-17 d8d9c6f
Rewrite software design section
