### 2026-06-17 a0f4b8d
Minor rewording
