### 2026-06-17 eaba9cf
Fix contribution link in README.md
Updated the contribution link to point to the main branch.
