### 2026-06-17 2535896
Update CONTRIBUTING.md
- Add link to GitHub profile
- Fix main branch name
