### 2026-06-17 aa9cae2
Add small section on financial support
