### 2026-06-17 12c48c2
Minor rewording of the software design section
