### 2026-07-06 #10485 @borgesaugusto
Hi! Thank you, @Redmar-van-den-Berg, for the clarification!
@bebatut or @samuelortion: Do you have any comments about the text added by the authors? Does it address your concerns about "software design" and "collaborative effort"?
Thank you!
