### 2026-08-25 #10485 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
