### 2026-08-25 #10485 @borgesaugusto
@Redmar-van-den-Berg Sorry, I didn't see it before. Please check the duplicated ORCID key in https://github.com/DCRT-LUMC/exonviz/blob/JOSS/paper.md?plain=1 (lines 8, 13, 18). That is probably the error being reported above.
