### 2026-08-25 #10485 @Redmar-van-den-Berg
@borgesaugusto I've removed the duplicate keys, I didn't notice before.
