### 2026-09-01 #10485 @Kevin-Mattheus-Moerman
👉 Please edit the ZENODO listed version to include the `v` so it ends up reading `Version v0.2.18`.

👉 Please spell out UK as United Kingdom
- [x] Copyedit pass/checked for typos
